Privacy and Accessibility updates
I believe it's important that users should be allowed to know what their data is actually used for. To this end, I've added a point to the 'Privacy' section about providing transparency regarding this.

I have also added the first point in an 'Accessibility' section. There is a real danger of developing countries being left behind by the information revolution, and we should strive to include them to the best of our abilities. Additionally, we should make software which is non-discriminatory. This means things like internationalisation should be considered properly, and assumptions about ability, sex, race or beliefs should be carefully considered.